Use the parametric background error instead od the ensemble std. dev. (#1960)
# Description
After a few days of digging into results and scratching our heads as to
why our tests have smaller o-b than the retros and @JohnSteffen-NOAA 's
hybrid ... We finally pin-pointed the issue to a typo in the hybrid B
configuration. This PR updates to a `jcb-gdas` hash that has the fix.

- fixes #1961 

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [x] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [x] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->