Bugfix for cice restart fix utility (#1974)
# Description

The cice restart fix utility on develop compares ice and snow volumes
(instead of thicknesses) to the threshold. This PR fixes it to compare
the thicknesses.

The behavior on develop leads to this script missing some of the worst
conditions. I have encountered a case, e.g., where we have ice
concentration of 0.00055 in one of ice categories and ice volume of 0.4
for that category, resulting in ~700m ice thickness in less than tenth
of percent of a gridpoint (I'd like to see that ice tower in real
life!). This gridpoint doesn't get picked up by the script as on
develop, but gets fixed with the update here.

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->