JCB-GDAS companion to GW obs staging and analysis stats refactor (#214)
This PR makes changes which support changes made in
https://github.com/NOAA-EMC/global-workflow/pull/4306 and
https://github.com/NOAA-EMC/GDASApp/pull/1999 in the following ways:
1. The obs statistics templates are moved from `algorithms/obsstats` to
`model/obsstats`, since these are indeed model templates and not
standalone algorithms.
2. The algorithm YAML for the JEDI obs stats application is moved from
the jcb-algorithms repo to this repo in `algorithms/obsstats`. The
justification is that the algorithm templates in jcb-algorithms should
be for general model-agnostics JEDI applications, usually existing in
the OOPS repo. This application on the other hand is specific to
GDASApp.
3. Minor variable names changes are made in the obs statistics
templates.