GDASApp companion for the GW obs staging and analysis stats refactor (#1999)
# Description

The companion PR supports changes made in the
https://github.com/NOAA-EMC/global-workflow/pull/4306 in the following
ways:
1. It consolidates the base yaml and JEDI config YAMLs for the analysis
stats job into a master config YAML and adds keys for the wxflow
FileHandler for staging and saving files for that job.
2. Any obs staging YAMLs are removed since that functionality is now
handled by the `Jedi` class in GW.
3. COM paths are removed from all JCB base YAMLs. Now, these paths are
passed as method arguments to the obs staging and diags saving methods
that now exist in the `Jedi` class. This way, the JCB base YAML is
strictly concerned with paths and file names for the run directories.
4. Changes are made to the JCB base YAMLs for the analysis stats job in
GW to be consistent with its refactoring in the companion PR.
5. An obs list is created for the analysis stats job, determining which
obs to create statistics for.

Additionally, `copy` and `link` keys in various YAMLs which go into the
FileHandler for several jobs become `copy_req` and `link_req`, since the
former is no deprecated.

# Companion PRs

https://github.com/NOAA-EMC/global-workflow/pull/4306
https://github.com/NOAA-EMC/jcb-gdas/pull/214
https://github.com/NOAA-EMC/jcb-algorithms/pull/17
https://github.com/NOAA-EMC/DA-utils/pull/49

# Issues

Refs https://github.com/NOAA-EMC/global-workflow/issues/4224
Refs https://github.com/NOAA-EMC/global-workflow/issues/4228

# Automated CI tests to run in Global Workflow

CI testing will be performed as part of review for the GW companion PR.
Preliminary testing has already been performed.